requirements.txt
# requirements.txt - STAR-Forecast 必要依赖

# ============ 核心依赖 ============
numpy>=1.24.0
pandas>=2.0.0

# ============ PyTorch 相关 ============
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# ============ 机器学习 ============
scikit-learn>=1.3.0
scipy>=1.11.0

# ============ 数据处理和可视化 ============
matplotlib>=3.7.0
seaborn>=0.12.0

# ============ 配置和日志 ============
pyyaml>=6.0
python-dotenv>=1.0.0
colorlog>=6.7.0

# ============ 工具库 ============
tqdm>=4.65.0
requests>=2.31.0

# ============ API 相关 ============
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# ============ 其他 ============
loguru>=0.7.0
rich>=13.7.0